+++ src/main.py
@@ -7,7 +7,7 @@ import logging
 import time
 
 from eldoria.app.app import main
-from eldoria.utils.logging import setup_logging
+from eldoria.app.logging import setup_logging
 
 if __name__ == "__main__":
     started_at = time.perf_counter()