+++ CHANGELOG.md
@@ -10,6 +10,7 @@ et le versioning suit [Semantic Versioning](https://semver.org/lang/fr/).
 ### Added
 - Ajout de la CI github-actions : setup pytest + docker build
 - Scripts pour automatiser les prochaines release
+- Ajout des fichiers de logs, récupérable via une commande admin
 
 ### Changed